Implement UDP functionality for video recording (UNTESTED)
Upon clicking the Record button on the website, the user will trigger the bbg to start recording the livestream footage. 30 seconds of footage will be recorded and the record button will be disabled for 45.

NOTE: This has not be tested. Testing will happen during the next team meeting where everyone has access to the webcam bbg.